@@ -1,17 +1,17 @@
 module gorm.io/sharding
 
-go 1.20
+go 1.21
 
 require (
 	github.com/bwmarrin/snowflake v0.3.0
 	github.com/longbridgeapp/assert v1.1.0
 	github.com/longbridgeapp/sqlparser v0.3.1
-	golang.org/x/exp v0.0.0-20230510235704-dd950f8aeaea
-	gorm.io/driver/mysql v1.4.7
-	gorm.io/driver/postgres v1.5.0
-	gorm.io/gorm v1.25.1
+	golang.org/x/exp v0.0.0-20230817173708-d852ddb80c63
+	gorm.io/driver/mysql v1.5.1
+	gorm.io/driver/postgres v1.5.2
+	gorm.io/gorm v1.25.4
 	gorm.io/hints v1.1.2
-	gorm.io/plugin/dbresolver v1.4.1
+	gorm.io/plugin/dbresolver v1.4.7
 )
 
 require (
@@ -19,12 +19,12 @@ require (
 	github.com/go-sql-driver/mysql v1.7.0 // indirect
 	github.com/jackc/pgpassfile v1.0.0 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
-	github.com/jackc/pgx/v5 v5.3.0 // indirect
+	github.com/jackc/pgx/v5 v5.3.1 // indirect
 	github.com/jinzhu/inflection v1.0.0 // indirect
 	github.com/jinzhu/now v1.1.5 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/stretchr/testify v1.8.1 // indirect
-	golang.org/x/crypto v0.6.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
+	golang.org/x/crypto v0.8.0 // indirect
+	golang.org/x/text v0.9.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
