@@ -22,7 +22,7 @@ func (pool ConnPool) PrepareContext(ctx context.Context, query string) (*sql.Stm
 	return pool.ConnPool.PrepareContext(ctx, query)
 }
 
-func (pool ConnPool) ExecContext(ctx context.Context, query string, args ...interface{}) (sql.Result, error) {
+func (pool ConnPool) ExecContext(ctx context.Context, query string, args ...any) (sql.Result, error) {
 	ftQuery, stQuery, table, err := pool.sharding.resolve(query, args...)
 	if err != nil {
 		return nil, err
@@ -42,7 +42,7 @@ func (pool ConnPool) ExecContext(ctx context.Context, query string, args ...inte
 }
 
 // https://github.com/go-gorm/gorm/blob/v1.21.11/callbacks/query.go#L18
-func (pool ConnPool) QueryContext(ctx context.Context, query string, args ...interface{}) (*sql.Rows, error) {
+func (pool ConnPool) QueryContext(ctx context.Context, query string, args ...any) (*sql.Rows, error) {
 	ftQuery, stQuery, table, err := pool.sharding.resolve(query, args...)
 	if err != nil {
 		return nil, err
@@ -61,7 +61,7 @@ func (pool ConnPool) QueryContext(ctx context.Context, query string, args ...int
 	return pool.ConnPool.QueryContext(ctx, stQuery, args...)
 }
 
-func (pool ConnPool) QueryRowContext(ctx context.Context, query string, args ...interface{}) *sql.Row {
+func (pool ConnPool) QueryRowContext(ctx context.Context, query string, args ...any) *sql.Row {
 	_, query, _, _ = pool.sharding.resolve(query, args...)
 	pool.sharding.querys.Store("last_query", query)
 
